A few more compiler warnings fixed.

+++ test/testautomation_sdltest.c
@@ -1161,7 +1161,7 @@ sdltest_randomAsciiStringWithMaximumLength(void *arg)
   size_t i;
 
   targetLen = 16 + SDLTest_RandomUint8();
-  result = SDLTest_RandomAsciiStringWithMaximumLength(targetLen);
+  result = SDLTest_RandomAsciiStringWithMaximumLength((int) targetLen);
   SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringWithMaximumLength(%d)", targetLen);
   SDLTest_AssertCheck(result != NULL, "Validate that result is not NULL");
   if (result != NULL) {
@@ -1182,7 +1182,7 @@ sdltest_randomAsciiStringWithMaximumLength(void *arg)
 
   /* Negative test */
   targetLen = 0;
-  result = SDLTest_RandomAsciiStringWithMaximumLength(targetLen);
+  result = SDLTest_RandomAsciiStringWithMaximumLength((int) targetLen);
   SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringWithMaximumLength(%d)", targetLen);
   SDLTest_AssertCheck(result == NULL, "Validate that result is NULL");
   lastError = (char *)SDL_GetError();
@@ -1215,7 +1215,7 @@ sdltest_randomAsciiStringOfSize(void *arg)
 
   /* Positive test */
   targetLen = 16 + SDLTest_RandomUint8();
-  result = SDLTest_RandomAsciiStringOfSize(targetLen);
+  result = SDLTest_RandomAsciiStringOfSize((int) targetLen);
   SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringOfSize(%d)", targetLen);
   SDLTest_AssertCheck(result != NULL, "Validate that result is not NULL");
   if (result != NULL) {
@@ -1236,7 +1236,7 @@ sdltest_randomAsciiStringOfSize(void *arg)
 
   /* Negative test */
   targetLen = 0;
-  result = SDLTest_RandomAsciiStringOfSize(targetLen);
+  result = SDLTest_RandomAsciiStringOfSize((int) targetLen);
   SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringOfSize(%d)", targetLen);
   SDLTest_AssertCheck(result == NULL, "Validate that result is NULL");
   lastError = (char *)SDL_GetError();
